apply changes from review
